requirements.txt
annotated-types==0.7.0
async-timeout==4.0.3
atomicwrites==1.4.1
attrs==22.2.0
beautifulsoup4==4.9.0
bidict==0.23.1
blinker==1.8.2
cachelib==0.13.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
eventlet==0.36.1
Flask==2.2.3
Flask-Cors==4.0.1
Flask-Login==0.6.2
Flask-Mail==0.10.0
Flask-Session==0.8.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.135.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==21.2.0
h11==0.14.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.1.2
more-itertools==9.0.0
msgspec==0.18.6
packaging==23.0
pluggy==0.13.1
proto-plus==1.24.0
protobuf==4.25.3
py==1.11.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
pyparsing==3.1.2
pytest==5.4.1
pytest-json-report==1.2.1
pytest-metadata==2.0.4
python-dotenv==0.13.0
python-engineio==4.9.1
python-socketio==5.11.3
redis==5.0.8
requests==2.32.3
rsa==4.9
simple-websocket==1.0.0
soupsieve==2.4
SQLAlchemy==2.0.7
tqdm==4.66.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.6
Werkzeug==2.2.3
wsproto==1.2.0
WTForms==3.0.1